Implementing JavaScript
Changed the page layout to have JavaScript at the top so items can be
added and written onto the page easily. Now the calcTotal function does
not work, but it should be very close. Will be adding more Javascript to
give the order confirmation on the page. Also, we were losing money on
hats, so I raised the cost :)