Branch Update
Added functional form to the merchandise page. Feedback is no longer in
an alert. Now feedback is color coated in the 'feedback' <div>. Also,
the order totaling is working good. Next is to add separate feedback
below the submit button. But a good step in the right direction with
this one :+1: