Final Push
Made the correct link active in the header when on my page. Also added
toFixed(2) to the orderTotal.
Added an entire form for the user to fill out, with some validation
logic, to fully simulate an order. Now the user sees the total, is
prompted to finish the steps, and then must enter appropriate data. Big
thanks to Anthony and Shane for being great online partners!